requirements.txt
Django==1.8.15
djangorestframework==3.5.1